+++ CHANGELOG.md
@@ -659,6 +659,7 @@
 - Sort the users list by name within each organisation
 - Only show Channel of delivery code for projects and third-party projects
 - Send approved report email notifications to BEIS users
+- Add "sdgs_apply" to CSV export
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-51...HEAD
 [release-51]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-50...release-51